Refactor: decouple ScriptVariables to \Setup\ProvideScriptVariables
